requirements.txt
autopep8
maturin
pillow
polars
slipcover
tqdm
unittest-xml-reporting
requests
